add polling of the user's currently playing song
Since the recently-played api only returns songs
that have finished, we have to also poll the user's
currently playing song. This adds some complexity
because the recently-played api tells you the
timestamp at which the playback finished, but
the currently-playing api tells you the timestamp
at which the playback started. Plays now also have
a played_at_minute timestamp that we can use to
compare the playback start time to the computed
(from playback_end - song_duration) playback start
time within a certain threshold to make a guess
as to whether that playback instance hasn't already
been captured by a recently-played sweep or a
currently-playing poll.